+++ tests/cpass.rs
@@ -2,7 +2,7 @@
 
 use heapless::{
     history_buf::HistoryBufView,
-    spsc::{Consumer, ConsumerView, Producer, ProducerView, Queue, QueueView},
+    spsc::{Consumer, Producer, Queue, QueueView},
     HistoryBuf, Vec, VecView,
 };
 
@@ -18,10 +18,8 @@ fn send() {
     {
     }
 
-    is_send::<Consumer<IsSend, 4>>();
-    is_send::<ConsumerView<IsSend>>();
-    is_send::<Producer<IsSend, 4>>();
-    is_send::<ProducerView<IsSend>>();
+    is_send::<Consumer<IsSend>>();
+    is_send::<Producer<IsSend>>();
     is_send::<Queue<IsSend, 4>>();
     is_send::<QueueView<IsSend>>();
     is_send::<Vec<IsSend, 4>>();