+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Made `LenType` opt-in.
 - Minor fixes to `pool::boxed` docs.
 - Add missing `Debug` derive to `vec::IntoIter`.
+- Removed generic from `spsc::Consumer`, `spsc::Producer` and `spsc::Iter`.
 
 ### Fixed